requirements.txt
fastapi
langchain
langchain-core
langchain-huggingface
pydantic
python-dotenv
uvicorn
